Extract querybuilder TS tests
